1、【优化】：人脸关键点算法优化，全面提升全脸区域准确性和稳定性2、【优化】：FUliveDemo初始化默认美颜效果优化，覆盖基础与通用美颜参数调节，提升默认美颜效果的适配性，满足更广泛用户的美颜需求。

+++ FUBeautyComponent/FUBeautyComponent/Model/FUBeautyFilterModel.h
@@ -13,7 +13,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 @property (nonatomic, assign) NSInteger filterIndex;
 @property (nonatomic, copy) NSString *filterName;
-@property (nonatomic, assign) double filterLevel;
+@property (nonatomic, assign) float filterLevel;
 
 @end
