+++ FUBeautyComponent/FUBeautyComponent/Model/FUBeautyShapeModel.h
@@ -17,9 +17,9 @@ NS_ASSUME_NONNULL_BEGIN
 /// 美型类型
 @property (nonatomic, assign) FUBeautyShape type;
 /// 当前值
-@property (nonatomic) double currentValue;
+@property (nonatomic) float currentValue;
 /// 默认值
-@property (nonatomic) double defaultValue;
+@property (nonatomic) float defaultValue;
 /// 默认值是否中位数
 @property (nonatomic, assign) BOOL defaultValueInMiddle;
 /// 设备性能等级要求
